prefetcher: ensure the groupTask.incompleteCount is allocated on a 64-bit aligned address (#3740)
## Summary

prefetcher: ensure the groupTask.incompleteCount is allocated on a 64-bit aligned address.

This is required in order to avoid the following on a ARM32:
```
panic: runtime error: invalid memory address or nil pointer dereference
[signal SIGSEGV: segmentation violation code=0x1 addr=0x0 pc=0xb541efb0]

goroutine 4220 [running]:
runtime/internal/atomic.goLoad64(0x8ae33be4, 0xb67e0fec, 0x8cf54500)
	runtime/internal/atomic/atomic_arm.go:131 +0x1c
github.com/algorand/go-algorand/ledger/internal.(*accountPrefetcher).prefetch(0x8cf54500, 0xb68023f0, 0x8cf532c0)
	github.com/algorand/go-algorand/ledger/internal/evalprefetcher.go:384 +0x1268
created by github.com/algorand/go-algorand/ledger/internal.prefetchAccounts
	github.com/algorand/go-algorand/ledger/internal/evalprefetcher.go:87 +0x128
```

## Test Plan

Test manually.

+++ ledger/internal/prefetcher/prefetcher_whitebox_test.go
@@ -24,7 +24,7 @@ func BenchmarkChannelWrites(b *testing.B) {
 	b.Run("groupTaskDone", func(b *testing.B) {
 		c := make(chan groupTaskDone, b.N)
 		for i := 0; i < b.N; i++ {
-			c <- groupTaskDone{groupIdx: i}
+			c <- groupTaskDone{groupIdx: int64(i)}
 		}
 	})
